DRY sort logic
